+++ tests/unit/files/datasets/test_get.py
@@ -1,7 +1,8 @@
 import re
 from unittest import TestCase, mock
 
-from zowe.zos_files_for_zowe_sdk import Files, exceptions, Datasets
+from zowe.zos_files_for_zowe_sdk import Datasets, Files, exceptions
+from zowe.zos_files_for_zowe_sdk.constants import ContentType
 
 
 class TestGetClass(TestCase):
@@ -22,7 +23,7 @@ class TestGetClass(TestCase):
         """Test list members sends request"""
         mock_send_request.return_value = mock.Mock(headers={"Content-Type": "application/json"}, status_code=200)
 
-        Files(self.test_profile).get_dsn_content(dataset_name="ds_name")
+        Files(self.test_profile).ds.retrieve_content(dataset_name="ds_name")
         mock_send_request.assert_called_once()
         prepared_request = mock_send_request.call_args[0][0]
         self.assertEqual(prepared_request.method, "GET")
@@ -32,7 +33,7 @@ class TestGetClass(TestCase):
         """Test list members sends request"""
         mock_send_request.return_value = mock.Mock(headers={"Content-Type": "application/json"}, status_code=200)
 
-        Files(self.test_profile).get_dsn_binary_content(dataset_name="ds_name")
+        Files(self.test_profile).ds.retrieve_content(dataset_name="ds_name", content_type=ContentType.BINARY)
         mock_send_request.assert_called_once()
         prepared_request = mock_send_request.call_args[0][0]
         self.assertEqual(prepared_request.method, "GET")