+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to the Zowe Client Python SDK will be documented in this fil
 
 ### Enhancements
 
+- Added `USSFiles.retrieve_content` as a substitution of `USSFiles.get_content` and `USSFiles.get_content_streamed`, `USSFiles.perform_download` as a substitution of `USSFiles.download`, `USSFiles.perform_upload` as a substitution of `USSFiles.upload`, `Datasets.retrieve_content` as a substitution of `Datasets.get_content` and `Datasets.get_binary_content`, `Datasets.perform_download` as a substitution of `Datasets.download` and `Datasets.download_binary`, `Datasets.perform_upload` as a substitution of `Datasets.upload`. [#380](https://github.com/zowe/zowe-client-python-sdk/pull/380)
+- Added `ContentType` enum class to distinguish between data types to submit / retrieve. [#380](https://github.com/zowe/zowe-client-python-sdk/pull/380)
 - Added binary USS files upload and write. [#380](https://github.com/zowe/zowe-client-python-sdk/pull/380)
 - Added `USSFiles.get_file_tag` function. [#380](https://github.com/zowe/zowe-client-python-sdk/pull/380)